Arreglado metodo registrar_pago

+++ club_sem_rust/lib.rs
@@ -498,17 +498,17 @@ mod club_sem_rust {
             let hoy = self.env().block_timestamp();
             if self.socios.len() > 0{
                 let i:i32 = 0;
-                while (i as usize) < self.socios.len() && self.socios[i].dni != dni{
+                while (i as usize) < self.socios.len() && self.socios[i as usize].dni != dni{
                      i = i + 1;
                 }
                 if self.socios[i].dni != dni{
                     panic!("EL DNI INGRESADO NO ES VALIDO");
                 } else{
-                    if self.socios[i].pagos[self.socios[i].pagos.len() - 1].pendiente = true {
-                        if self.socios[i].cumple_bonificacion(){
-                            self.socios[i].pagos[self.socios[i].pagos.len() - 1].realizar_pago(Some(self.descuento), monto, hoy, self.precio_categorias);
+                    if self.socios[i as usize].pagos[self.socios[i as usize].pagos.len() - 1].pendiente = true {
+                        if self.socios[i as usize].cumple_bonificacion(){
+                            self.socios[i as usize].pagos[self.socios[i as usize].pagos.len() - 1].realizar_pago(Some(self.descuento), monto, hoy, self.precio_categorias);
                         }else{
-                            self.socios[i].pagos[self.socios[i].pagos.len() - 1].realizar_pago(None, monto, hoy, self.precio_categorias);
+                            self.socios[i as usize].pagos[self.socios[i as usize].pagos.len() - 1].realizar_pago(None, monto, hoy, self.precio_categorias);
                         }
                         
                     }else{
